requirements.txt
Django==4.2.23
psycopg2-binary
soft-delete==0.2.2
django-simple-history==3.7.0
django-localflavor==4.0
django-recaptcha==4.0.0
django-select2==8.2.1
django_widget_tweaks==1.5.0
ibge==0.0.5
pillow==10.4.0